+++ aquadoggo/src/rpc/api.rs
@@ -22,55 +22,3 @@ pub fn rpc_api_handler(pool: Pool) -> RpcApiService {
         .with_method("panda_queryEntries", query_entries)
         .finish()
 }
-
-#[cfg(test)]
-mod tests {
-    use tide_testing::TideTestingExt;
-
-    use crate::rpc::server::build_rpc_server;
-    use crate::test_helpers::{initialize_db, random_entry_hash, rpc_error, rpc_request};
-
-    use super::rpc_api_handler;
-
-    #[async_std::test]
-    async fn respond_with_method_not_allowed() {
-        let pool = initialize_db().await;
-        let rpc_api_handler = rpc_api_handler(pool.clone());
-        let app = build_rpc_server(rpc_api_handler);
-
-        assert_eq!(
-            app.get("/").recv_string().await.unwrap(),
-            "Used HTTP Method is not allowed. POST or OPTIONS is required"
-        );
-    }
-
-    #[async_std::test]
-    async fn respond_with_wrong_author_error() {
-        let pool = initialize_db().await;
-        let rpc_api_handler = rpc_api_handler(pool.clone());
-        let app = build_rpc_server(rpc_api_handler);
-
-        let request = rpc_request(
-            "panda_getEntryArguments",
-            &format!(
-                r#"{{
-                    "author": "1234",
-                    "schema": "{}"
-                }}"#,
-                random_entry_hash()
-            ),
-        );
-
-        let response = rpc_error("invalid author key length");
-
-        let response_body: serde_json::value::Value = app
-            .post("/")
-            .body(tide::Body::from_string(request.into()))
-            .content_type("application/json")
-            .recv_json()
-            .await
-            .unwrap();
-
-        assert_eq!(response_body.to_string(), response);
-    }
-}
