+++ aquadoggo/src/rpc/server.rs
@@ -71,7 +71,6 @@ pub fn build_rpc_server(api: RpcApiService) -> RpcServer {
         .with(WebSocket::new(handle_ws_request))
         .get(|_| async { Ok("Used HTTP Method is not allowed. POST or OPTIONS is required") })
         .post(handle_http_request);
-
     app
 }
 
@@ -81,4 +80,25 @@ pub async fn start_rpc_server(config: &Configuration, api: RpcApiService) -> any
     let server = build_rpc_server(api);
     server.listen(http_address).await?;
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use tide_testing::TideTestingExt;
+
+    use crate::rpc::api::rpc_api_handler;
+    use crate::rpc::server::build_rpc_server;
+    use crate::test_helpers::initialize_db;
+
+    #[async_std::test]
+    async fn respond_with_method_not_allowed() {
+        let pool = initialize_db().await;
+        let rpc_api_handler = rpc_api_handler(pool.clone());
+        let app = build_rpc_server(rpc_api_handler);
+
+        assert_eq!(
+            app.get("/").recv_string().await.unwrap(),
+            "Used HTTP Method is not allowed. POST or OPTIONS is required"
+        );
+    }
 }
